Rename script file
